Correct usage of cookbook name
@edPratt, my apologies for missing this, but the name of the cookbook
should be `et_apache_test`; the `-cookbook` appendage in the repo name
is just a convention we use at EverTrue. No need to include it in the
cookbook name, since, well, we all know it’s a cookbook by the time
you’re actually calling it by name in recipes & run lists.